+++ src/ui_bevy/mod.rs
@@ -1,7 +1,5 @@
 mod plugin;
-mod ui_builder_react_ext;
 mod ui_ext;
 
 pub(crate) use plugin::*;
-pub use ui_builder_react_ext::*;
 pub use ui_ext::*;